+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Categories for each release: Added, Changed, Deprecated, Removed, Fixed, Securit
 
 ### Added
 
+* Support for Preserve Job Outputs in `dx run` using `--preserve-job-outputs` and `--preserve-job-outputs-folder`
 * Pretty-printing Preserve Job Outputs Folder field in `dx describe` for jobs and analyses
 
 ## [343.0] - beta